+++ src/RcppExports.cpp
@@ -6,66 +6,6 @@
 
 using namespace Rcpp;
 
-// EntropyUT
-double EntropyUT(int G, arma::vec pro, arma::mat mean, arma::cube sigma, double d);
-RcppExport SEXP _ppgmmga_EntropyUT(SEXP GSEXP, SEXP proSEXP, SEXP meanSEXP, SEXP sigmaSEXP, SEXP dSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< int >::type G(GSEXP);
-    Rcpp::traits::input_parameter< arma::vec >::type pro(proSEXP);
-    Rcpp::traits::input_parameter< arma::mat >::type mean(meanSEXP);
-    Rcpp::traits::input_parameter< arma::cube >::type sigma(sigmaSEXP);
-    Rcpp::traits::input_parameter< double >::type d(dSEXP);
-    rcpp_result_gen = Rcpp::wrap(EntropyUT(G, pro, mean, sigma, d));
-    return rcpp_result_gen;
-END_RCPP
-}
-// EntropyVAR
-double EntropyVAR(int G, NumericVector pro, NumericMatrix mean, arma::cube sigma, int d);
-RcppExport SEXP _ppgmmga_EntropyVAR(SEXP GSEXP, SEXP proSEXP, SEXP meanSEXP, SEXP sigmaSEXP, SEXP dSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< int >::type G(GSEXP);
-    Rcpp::traits::input_parameter< NumericVector >::type pro(proSEXP);
-    Rcpp::traits::input_parameter< NumericMatrix >::type mean(meanSEXP);
-    Rcpp::traits::input_parameter< arma::cube >::type sigma(sigmaSEXP);
-    Rcpp::traits::input_parameter< int >::type d(dSEXP);
-    rcpp_result_gen = Rcpp::wrap(EntropyVAR(G, pro, mean, sigma, d));
-    return rcpp_result_gen;
-END_RCPP
-}
-// EntropySOTE
-double EntropySOTE(arma::mat data, int G, arma::vec pro, arma::mat mean, arma::cube sigma);
-RcppExport SEXP _ppgmmga_EntropySOTE(SEXP dataSEXP, SEXP GSEXP, SEXP proSEXP, SEXP meanSEXP, SEXP sigmaSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< arma::mat >::type data(dataSEXP);
-    Rcpp::traits::input_parameter< int >::type G(GSEXP);
-    Rcpp::traits::input_parameter< arma::vec >::type pro(proSEXP);
-    Rcpp::traits::input_parameter< arma::mat >::type mean(meanSEXP);
-    Rcpp::traits::input_parameter< arma::cube >::type sigma(sigmaSEXP);
-    rcpp_result_gen = Rcpp::wrap(EntropySOTE(data, G, pro, mean, sigma));
-    return rcpp_result_gen;
-END_RCPP
-}
-// EntropyMCapprox
-List EntropyMCapprox(arma::mat data, int G, arma::vec pro, arma::mat mean, arma::cube sigma);
-RcppExport SEXP _ppgmmga_EntropyMCapprox(SEXP dataSEXP, SEXP GSEXP, SEXP proSEXP, SEXP meanSEXP, SEXP sigmaSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< arma::mat >::type data(dataSEXP);
-    Rcpp::traits::input_parameter< int >::type G(GSEXP);
-    Rcpp::traits::input_parameter< arma::vec >::type pro(proSEXP);
-    Rcpp::traits::input_parameter< arma::mat >::type mean(meanSEXP);
-    Rcpp::traits::input_parameter< arma::cube >::type sigma(sigmaSEXP);
-    rcpp_result_gen = Rcpp::wrap(EntropyMCapprox(data, G, pro, mean, sigma));
-    return rcpp_result_gen;
-END_RCPP
-}
 // LinTransf
 List LinTransf(arma::mat mean, arma::cube sigma, arma::mat B, arma::mat Z, int G, int d);
 RcppExport SEXP _ppgmmga_LinTransf(SEXP meanSEXP, SEXP sigmaSEXP, SEXP BSEXP, SEXP ZSEXP, SEXP GSEXP, SEXP dSEXP) {
@@ -128,6 +68,91 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< int >::type d(dSEXP);
     Rcpp::traits::input_parameter< int >::type p(pSEXP);
     rcpp_result_gen = Rcpp::wrap(encodebasis(par, d, p));
+    return rcpp_result_gen;
+END_RCPP
+}
+// logsumexp
+double logsumexp(arma::vec x);
+RcppExport SEXP _ppgmmga_logsumexp(SEXP xSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::vec >::type x(xSEXP);
+    rcpp_result_gen = Rcpp::wrap(logsumexp(x));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EntropyUT
+double EntropyUT(int G, arma::vec pro, arma::mat mean, arma::cube sigma, double d);
+RcppExport SEXP _ppgmmga_EntropyUT(SEXP GSEXP, SEXP proSEXP, SEXP meanSEXP, SEXP sigmaSEXP, SEXP dSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< int >::type G(GSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type pro(proSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type mean(meanSEXP);
+    Rcpp::traits::input_parameter< arma::cube >::type sigma(sigmaSEXP);
+    Rcpp::traits::input_parameter< double >::type d(dSEXP);
+    rcpp_result_gen = Rcpp::wrap(EntropyUT(G, pro, mean, sigma, d));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EntropyVAR
+double EntropyVAR(int G, NumericVector pro, NumericMatrix mean, arma::cube sigma, int d);
+RcppExport SEXP _ppgmmga_EntropyVAR(SEXP GSEXP, SEXP proSEXP, SEXP meanSEXP, SEXP sigmaSEXP, SEXP dSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< int >::type G(GSEXP);
+    Rcpp::traits::input_parameter< NumericVector >::type pro(proSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type mean(meanSEXP);
+    Rcpp::traits::input_parameter< arma::cube >::type sigma(sigmaSEXP);
+    Rcpp::traits::input_parameter< int >::type d(dSEXP);
+    rcpp_result_gen = Rcpp::wrap(EntropyVAR(G, pro, mean, sigma, d));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EntropySOTE
+double EntropySOTE(arma::mat data, int G, arma::vec pro, arma::mat mean, arma::cube sigma);
+RcppExport SEXP _ppgmmga_EntropySOTE(SEXP dataSEXP, SEXP GSEXP, SEXP proSEXP, SEXP meanSEXP, SEXP sigmaSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat >::type data(dataSEXP);
+    Rcpp::traits::input_parameter< int >::type G(GSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type pro(proSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type mean(meanSEXP);
+    Rcpp::traits::input_parameter< arma::cube >::type sigma(sigmaSEXP);
+    rcpp_result_gen = Rcpp::wrap(EntropySOTE(data, G, pro, mean, sigma));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EntropyMCapprox
+List EntropyMCapprox(arma::mat data, int G, arma::vec pro, arma::mat mean, arma::cube sigma);
+RcppExport SEXP _ppgmmga_EntropyMCapprox(SEXP dataSEXP, SEXP GSEXP, SEXP proSEXP, SEXP meanSEXP, SEXP sigmaSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat >::type data(dataSEXP);
+    Rcpp::traits::input_parameter< int >::type G(GSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type pro(proSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type mean(meanSEXP);
+    Rcpp::traits::input_parameter< arma::cube >::type sigma(sigmaSEXP);
+    rcpp_result_gen = Rcpp::wrap(EntropyMCapprox(data, G, pro, mean, sigma));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EntropyGMM
+double EntropyGMM(arma::mat data, arma::vec pro, arma::mat mean, arma::cube sigma);
+RcppExport SEXP _ppgmmga_EntropyGMM(SEXP dataSEXP, SEXP proSEXP, SEXP meanSEXP, SEXP sigmaSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat >::type data(dataSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type pro(proSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type mean(meanSEXP);
+    Rcpp::traits::input_parameter< arma::cube >::type sigma(sigmaSEXP);
+    rcpp_result_gen = Rcpp::wrap(EntropyGMM(data, pro, mean, sigma));
     return rcpp_result_gen;
 END_RCPP
 }