+++ NEWS.md
@@ -1,5 +1,7 @@
-# ppgmmga 1.2 (2019-07)
+# ppgmmga 1.2 (2019-10)
 
+- Added closed-form formula for computing the Negentropy of GMMs. 
+- Changed default in ppgmmga() function call to use approx = "none", i.e. to use closed-form formula for computing the Negentropy.
 - Updated references to JCGS paper.
 - Added examples in ppgmmaga.Rd and vignette of possible usage of spinplot() from 'msir' package.
 - Updated vignette.